% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{aov.pheno.test}
\alias{aov.pheno.test}
\title{anova test for phenotype based on haplotype}
\usage{
aov.pheno.test(haps, gene, feature)
}
\arguments{
\item{haps}{SeqHap object}

\item{gene}{gene name}

\item{feature}{pheaotype name}
}
\description{
anova test for phenotype based on haplotype
}
\author{
Kai Guo
}
