% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.colnames.mcol}
\alias{.colnames.mcol}
\title{change mcols data colnames}
\usage{
.colnames.mcol(gr, names)
}
\arguments{
\item{gr}{GRanges object}

\item{names}{new column names}
}
\description{
change mcols data colnames
}
\author{
Kai Guo
}
