% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.detect_file}
\alias{.detect_file}
\title{detect file format for import gff or gff3 file
modified based on GenomicFeatures .detect_file_format function}
\usage{
.detect_file(file)
}
\description{
detect file format for import gff or gff3 file
modified based on GenomicFeatures .detect_file_format function
}
