% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{.GENE_TYPES}
\alias{.GENE_TYPES}
\title{copy from GenomicFeatures}
\format{
An object of class \code{character} of length 3.
}
\usage{
.GENE_TYPES
}
\description{
copy from GenomicFeatures
}
\keyword{datasets}
