% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expandRange}
\alias{expandRange}
\title{expand range from start and end}
\usage{
expandRange(gr, upstream = 1000, downstream = 1000)
}
\arguments{
\item{gr}{GRanges object}

\item{upstream}{flanking length for upstream}

\item{downstream}{flanking length for downstream}
}
\description{
expand range from start and end
}
\author{
Kai Guo
}
