% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{findover}
\alias{findover}
\title{find overlap between target genes and hmp}
\usage{
findover(gr, hmp, upstream = 2000, downstream = 500)
}
\arguments{
\item{gr}{GrangesInfo object or GrangeList}

\item{hmp}{data.frame include hapmap information}

\item{upstream}{Vectors of non-NA non-negative integers.
Defines the number of nucleotides toward the 5'}

\item{downstream}{Vectors of non-NA non-negative integers.
Defines the number toward the 3'end,relative to the transcription start site}
}
\value{
OverlapInfo object
}
\description{
find overlap between target genes and hmp
}
\author{
Kai Guo
}
