% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hapnet}
\alias{hapnet}
\title{plot haplotype with network}
\usage{
hapnet(
  hap,
  gene,
  feature = NULL,
  freq = TRUE,
  cex = 1,
  altlinks = FALSE,
  high = "red",
  low = "skyblue",
  log = TRUE,
  edge.label.size = 3,
  node.label.size = 4,
  node.alpha = 0.75,
  ...
)
}
\description{
plot haplotype with network
}
