% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{importGFF}
\alias{importGFF}
\title{read GFF file into Granges object}
\usage{
importGFF(file, format = c("auto", "gtf", "gff3"))
}
\arguments{
\item{file}{filename for gtf or gff3}

\item{format}{character indicate the file format(gtf/gff3)}
}
\value{
GRanges object
}
\description{
import gff3 or gtf file
}
\author{
Kai Guo
}
