% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OverlapInfo-class}
\alias{OverlapInfo-class}
\title{Class "OverlapInfo"
This class represents the extract gene infomation from}
\description{
Class "OverlapInfo"
This class represents the extract gene infomation from
}
\section{Slots}{

\describe{
\item{\code{gr}}{GrangesList of gff or gff3 files with filter genes}

\item{\code{overlap}}{GRangesList}

\item{\code{gene}}{vector of character of genes}

\item{\code{gene}}{Gene IDs}

\item{\code{format}}{a character gff/gtf or gff3}

\item{\code{mcol}}{a character specify the gene id column}

\item{\code{cols}}{a vector include gene structure information}
}}

\author{
Kai Guo
}
\keyword{classes}
