% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{preGranges}
\alias{preGranges}
\title{filter GRanges and add flanking regions}
\usage{
preGranges(
  gr,
  gene = NULL,
  exon = TRUE,
  cds = FALSE,
  gene_name = FALSE,
  format = NULL
)
}
\arguments{
\item{gr}{GRanges object}

\item{gene}{gene name}

\item{exon}{a logical value indicating whether use exon information or not}

\item{cds}{a logical value indicating whether use cds information or not}

\item{gene_name}{a logical value indicating whether use
gene_name or not(for gtf)}

\item{format}{If not missing, should be one of “gff”,“gff3” or “gtf”.}
}
\value{
GRangesInfo object include only select genes information
}
\description{
filter GRanges and add flanking regions
}
\author{
Kai Guo
}
