% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{read.data}
\alias{read.data}
\title{prepare GWAS data}
\usage{
read.data(file, sep = "\\t", comment = "&")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{sep}{the field separator character.}

\item{comment}{character: a character vector of length one containing a
single character or an empty string.}
}
\value{
GRanges
}
\description{
prepare GWAS data
}
\author{
Kai Guo
}
