% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read.vcf}
\alias{read.vcf}
\title{read vcf files}
\usage{
read.vcf(file, genome = "hg19", param = ScanVcfParam())
}
\arguments{
\item{file}{A ‘VcfFile’ (synonymous with ‘TabixFile’) instance or
character() name of the VCF file to be processed. When ranges
are specified in ‘param’, ‘file’ must be a ‘VcfFile’.}

\item{genome}{A ‘character’ or ‘Seqinfo’ object.}

\item{param}{param: An instance of ‘ScanVcfParam’, ‘GRanges’, ‘GRangesList’ or
‘IntegerRangesList’}
}
\value{
GRanges object
}
\description{
read vcf files
}
\author{
Kai Guo
}
