% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_pheno}
\alias{read_pheno}
\title{read hmp files and combine snp together}
\usage{
read_pheno(file, sep = "\\t")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{sep}{the field separator character.}

\item{comment}{character: a character vector of length one containing a
single character or an empty string.}
}
\value{
data.frame
}
\description{
read hmp files and combine snp together
}
\author{
Kai Guo
}
