% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_vcf}
\alias{read_vcf}
\title{read vcf files}
\usage{
read_vcf(file)
}
\arguments{
\item{file}{A ‘VcfFile’ (synonymous with ‘TabixFile’) instance or
character() name of the VCF file to be processed..}
}
\value{
GRanges object
}
\description{
read vcf files
}
\author{
Kai Guo
}
