% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{results,SeqHap-method}
\alias{results,SeqHap-method}
\title{extract final results from SeqHap object}
\usage{
\S4method{results}{SeqHap}(object, gene = NULL)
}
\arguments{
\item{object}{SeqHap object}

\item{gene}{gene}
}
\description{
extract final results from SeqHap object
}
\author{
Kai Guo
}
