% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeqHap-class}
\alias{SeqHap-class}
\title{Class "SeqHap"
This class represents the result of overlap between snp and target gene}
\description{
Class "SeqHap"
This class represents the result of overlap between snp and target gene
}
\section{Slots}{

\describe{
\item{\code{gr}}{GRangeList obejct}

\item{\code{pheno}}{data frame of phenotype data}

\item{\code{haplotype}}{list of haplotype object}

\item{\code{gene}}{vector of character of genes}
}}

\author{
Kai Guo
}
\keyword{classes}
