% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{seqs,SeqHap-method}
\alias{seqs,SeqHap-method}
\title{extract sequence form SeqHap object}
\usage{
\S4method{seqs}{SeqHap}(object, gene = NULL)
}
\arguments{
\item{object}{SeqHap object}

\item{gene}{gene}
}
\description{
extract sequence form SeqHap object
}
\author{
Kai Guo
}
