% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{snboxplot}
\alias{snboxplot}
\title{boxplot for haplotype testing based on phenotype}
\usage{
snboxplot(hap, gene, feature)
}
\arguments{
\item{hap}{SeqHap object include a GRanges, GRangesList}

\item{gene}{gene name}

\item{feature}{phenotype name}
}
\description{
boxplot for haplotype testing based on phenotype
}
\author{
Kai Guo
}
