% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp}
\alias{snp}
\title{snp}
\format{
A list with three data.frame:
\describe{
\item{gff}{GRanges format of gff3}
\item{hap}{hapmap data}
\item{pheno}{phenotype data}
}
}
\usage{
snp
}
\description{
snp include gff3,hapmap and phenotype information
}
\examples{
snp

}
\keyword{datasets}
