% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{snp2fasta}
\alias{snp2fasta}
\title{snp data frame to vector}
\usage{
snp2fasta(snp)
}
\arguments{
\item{snp}{data.frame with snp information}
}
\value{
sequence file
}
\description{
snp data frame to vector
}
\author{
Kai Guo
}
