% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{snp2hap}
\alias{snp2hap}
\title{extract fasta from genotype file and find haplotype}
\usage{
snp2hap(file, grob)
}
\arguments{
\item{file}{phenotype file}

\item{grob}{GRanges of overlap results}
}
\value{
SeqHap object
}
\description{
extract fasta from genotype file and find haplotype
}
\author{
Kai Guo
}
