% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{snplot}
\alias{snplot}
\title{plot snp with gene}
\usage{
snplot(
  hap,
  gene,
  hapname = NULL,
  layout = "lolliplot",
  snp.colors = NULL,
  angle = 90,
  side = FALSE,
  type = "circle",
  rescale = FALSE,
  snp.alpha = 0.7,
  jitter = "node",
  random = TRUE,
  snp.lwd = 0.5,
  label.cex = 0.8,
  upstream = NULL,
  downstream = NULL
)
}
\arguments{
\item{hap}{SeqHap object include a GRanges, GRangesList}

\item{gene}{character gene}

\item{hapname}{haplotype want to display}

\item{layout}{a character to indiate layout method (lolliplot or
dandelion.plot)}

\item{snp.colors}{color for the snp, random colors will be used if NULL}

\item{angle}{a numeric indicate the angle for the labels}

\item{side}{use both top and bottom to display the SNP (TRUE/FALSE)}

\item{type}{character. Could be circle, pie, pin, pie.stack or flag.}

\item{rescale}{Recalse the x-axis or not.}

\item{snp.alpha}{alpha value for color}

\item{jitter}{jitter the position of nodes or labels.}

\item{random}{use random height to display SNP or not}

\item{snp.lwd}{The line width for displaying SNP, a _positive_ number.}

\item{label.cex}{label size}

\item{upstream}{Defines the number of nucleotides toward the 5'}

\item{downstream}{Defines the number of nucleotides toward the 3'}
}
\value{
figure
}
\description{
plot snp with gene
}
\author{
Kai Guo
}
