% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetrack.R
\name{snptrack}
\alias{snptrack}
\title{plot Gene Track with SNP information}
\usage{
snptrack(
  obj,
  dat,
  id = "gene_id",
  color = NULL,
  chr = NULL,
  region = NULL,
  allchr = FALSE,
  gene = NULL,
  threshold = -log10(1e-06),
  geneOnly = FALSE,
  exonOnly = FALSE,
  upstream = 1000,
  downstream = 1000,
  alpha.point = 0.75,
  point.size = 1,
  point.shape = 20,
  exon = "darkgreen",
  utr3 = "cyan4",
  utr5 = "cyan4",
  intron = "black",
  gene.label.size = 0.5,
  high = "cyan4",
  low = "lightblue",
  threshold.col = "red",
  threshold.lwd = 1,
  threshold.lty = 2,
  ylab = "-log10 (P value)",
  legend.lab = NULL,
  ylab.size = 0.9,
  arrow.col = "lightblue",
  arrow.fill = "lightblue"
)
}
\arguments{
\item{obj}{Granges object include gff file}

\item{dat}{Granges object include GWAS results}

\item{color}{color for the point}

\item{chr}{chromosome name}

\item{region}{a vector include the start and the end location}

\item{allchr}{choose TRUE when displaying whole chromosome or large region}

\item{gene}{gene name}

\item{threshold}{threshold to add the line}

\item{geneOnly}{only show points in the gene region}

\item{exonOnly}{only show points in the exon region}

\item{upstream}{upstream bp}

\item{downstream}{downstream bp}

\item{alpha.point}{alpha}

\item{point.size}{size of the point}

\item{point.shape}{shape of the point}

\item{exon}{exon color}

\item{utr3}{3'utr color}

\item{utr5}{5'utr color}

\item{intron}{intron color}

\item{gene.label.size}{size of the gene label}

\item{high}{legend color}

\item{low}{low legend color}

\item{threshold.col}{color for threshold line}

\item{threshold.lwd}{line width for threshold line}

\item{threshold.lty}{line type for threshold line}

\item{ylab}{ylab name}

\item{ylab.size}{ylab size}

\item{arrow.col}{arrow color}

\item{arrow.fill}{arrow fill}
}
\description{
plot Gene Track with SNP information
}
\author{
Kai Guo
}
